moved all methods from game to util
